@@ -1,8 +1,7 @@
 import React, { useEffect } from "react";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { Home } from "./pages/Home/Home";
 import { NotFound } from "./pages/404/NotFound";
-import crypto from "crypto";
 import { Footer } from "./components/footer/footer";
 import { Orders } from "./pages/Orders/Orders";
 
@@ -10,7 +9,7 @@ const App = (): React.ReactElement => {
 	useEffect(() => {
 		const userId = localStorage.getItem("userId");
 		if (!userId) {
-			const id = crypto.randomBytes(20).toString("hex");
+			const id = crypto.randomUUID();
 			localStorage.setItem("userId", id);
 		}
 	}, []);
@@ -18,11 +17,11 @@ const App = (): React.ReactElement => {
 	return (
 		<Router>
 			<div className="app-container">
-				<Switch>
-					<Route exact path="/orders" component={Orders} />
-					<Route exact path="/" component={Home} />
-					<Route component={NotFound} />
-				</Switch>
+				<Routes>
+					<Route path="/orders" element={<Orders />} />
+					<Route path="/" element={<Home />} />
+					<Route element={<NotFound />} />
+				</Routes>
 				<Footer />
 			</div>
 		</Router>
