@@ -1,11 +1,9 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App";
 
-ReactDOM.render(
-	<React.StrictMode>
-		<App />
-	</React.StrictMode>,
-	document.getElementById("root")
-);
+const container = document.getElementById("root");
+const root = createRoot(container as HTMLElement); // createRoot(container!) if you use TypeScript
+
+root.render(<App />);
