@@ -1,16 +1,15 @@
 import React from "react";
-import { RouteComponentProps } from "react-router";
 import styles from "./Home.module.scss";
 import { HeaderBackground } from "./HeaderBackground";
 import { TopPizzas } from "./TopPizzas";
 import { PizzasMenu } from "./PizzasMenu";
 import { NavBar } from "../../components/navbar/navbar";
 
-export const Home = ({ history }: RouteComponentProps): React.ReactElement => {
+export const Home = (): React.ReactElement => {
 	return (
 		<div className={styles.homeContainer}>
 			<NavBar />
-			<HeaderBackground history={history} />
+			<HeaderBackground />
 			<TopPizzas />
 			<PizzasMenu />
 		</div>
